pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "skill-jepa"
version = "0.1.0"
description = "A Predictive World Model over Latent Action Manifolds (Macro-Steps)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Tomás Couso Coddou", email = "tomascousocoddou@gmail.com"}
]

dependencies = [
    # Core Machine Learning
    "torch>=2.0.0",
    "torchvision",
    "pytorch-lightning",
    "einops",          # Crucial for JEPA tensor routing
    
    # Utilities & Configuration
    "numpy",
    "scipy",
    "scikit-learn",    # Required for SWM normalizers (StandardScaler)
    "hydra-core",      # Required for the eval.py solver instantiation
    "omegaconf",       # Configuration management for Hydra
    "pyyaml",
    
    # Data & Logging
    "h5py",            # Core backend for stable-worldmodel
    "wandb",
    "tqdm",
    "pillow",
    "matplotlib"
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pytest",
    "jupyter"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"